add hook for use admin-pages
